Rework elision
